Add -s option to allow changing the initially selected date
This is useful in scripts where datepicker is invoked more than once,
for example, to input a date range. In this case, datepicker has to be
started twice to obtain both a start and an end date. When spawning
datepicker the second time to request the end date, it is often
desirable to have the previously obtained start date be the initially
selected date for the end date selection.